correct handling of `history` args
This fixes several problems with how the builtin `history` command handles
arguments. It now complains and refuses to do anything if the user specifies
incompatible actions (e.g., `--search` and `--clear`). It also fixes a
regression introduced by previous changes with regard to invocations that
don't explicitly specify `--search` or a search term.

This change is already far larger than I like so unit tests will be added
in a separate commit.

Partially addresses issue #3224. It does not fix the problems with the
interactive `history --delete` command in the `history` fish script. That will
be dealt with via issue #31.