make shadowing a builtin a warning
Further discussion has lead to the decision that defining a function that
shadows a builtin without using `--shadow-builtin` should not keep the
function from being created. It should still print a warning to stderr
so the user knows there may be problems. This also improves the wording
of the respective error messages.

Fixes #3319