Fix print make targets
Suppress the output both to stdout and stderr in the check for the
option. Also provide the directory (`C`) instead of utilizing an empty
Makefile.

It seems the empty Makefile (`/dev/null`) returns status code 1 which
fails the check and means that it will always be assumed to be FreeBSD.