Fix string regex replace hang up
If a parse error occurred in read_unquoted_escape(), it returns 0 to
the caller. But the result would be used as an increment in function
interpret_escapes() without any check, leading to an infinite loop.
The fix here checks the return value and jumps out of the loop if we
get 0.