allow configuring how undef vars are handled
This introduces the `fish_undef_var` variable. Set it to

 -1 : allow undef vars (fish 2.x behavior)
  0 : allow undef vars but warn when they are seen
  1 : disallow undef vars except in `test` commands
  2 : disallow undef vars in all situations

This does not provide context for the fish_undef_var=0 case. That's
because doing so requires more invasive changes such as adding another
symbol to `enum expand_error_t`. And doing so is hard to justify since
we don't want that capability to be present in the final fish release.

Fixes #4163