Better (faster) implementation of unique_unsorted
I forgot that we are using wcstring, not char *, which means that we can
use std::hash<T>(..) instead of requiring an outside string hashing
library.

This code is both faster and more memory-friendly. If acceptable, we can
switch to using std::unordered_set<T> instead of std::set<T> to store
the seen hashes.